requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-arm64
absl-py=1.3.0=pypi_0
aom=3.5.0=h7ea286d_0
astunparse=1.6.3=pypi_0
atk-1.0=2.36.0=h7fe96df_0
brotli=1.0.9=h1c322ee_7
brotli-bin=1.0.9=h1c322ee_7
brotlipy=0.7.0=py310hf8d0d8f_1004
bzip2=1.0.8=h3422bc3_4
c-ares=1.18.1=h3422bc3_0
ca-certificates=2022.9.24=h4653dfc_0
cached-property=1.5.2=hd8ed1ab_1
cached_property=1.5.2=pyha770c72_1
cachetools=5.2.0=pypi_0
cairo=1.16.0=h29d4eff_2
certifi=2022.9.24=pyhd8ed1ab_0
cffi=1.15.1=py310he00a5c5_0
charset-normalizer=2.1.1=pyhd8ed1ab_0
colorama=0.4.5=pyhd8ed1ab_0
conda=22.9.0=py310hbe9552e_2
conda-package-handling=1.8.1=py310hf8d0d8f_1
contourpy=1.0.5=py310h2887b22_0
cryptography=37.0.4=py310h94bb23d_0
cycler=0.11.0=pyhd8ed1ab_0
expat=2.4.9=hb7217d7_0
ffmpeg=5.1.2=gpl_hcccd11d_101
flatbuffers=22.9.24=pypi_0
font-ttf-dejavu-sans-mono=2.37=hab24e00_0
font-ttf-inconsolata=3.000=h77eed37_0
font-ttf-source-code-pro=2.038=h77eed37_0
font-ttf-ubuntu=0.83=hab24e00_0
fontconfig=2.14.0=h82840c6_1
fonts-conda-ecosystem=1=0
fonts-conda-forge=1=0
fonttools=4.37.4=py310h8e9501a_0
freetype=2.12.1=hd633e50_0
fribidi=1.0.10=h1a28f6b_0
gast=0.4.0=pypi_0
gdk-pixbuf=2.42.6=h74ef11a_4
gettext=0.21.1=h0186832_0
giflib=5.2.1=h1a28f6b_0
glib=2.69.1=h98b2900_1
gmp=6.2.1=h9f76cd9_0
gnutls=3.7.8=h9f1a10d_0
gobject-introspection=1.68.0=py310h3459c9f_3
google-auth=2.13.0=pypi_0
google-auth-oauthlib=0.4.6=pypi_0
google-pasta=0.2.0=pypi_0
graphite2=1.3.14=hc377ac9_1
graphviz=2.48.0=h6d4e08e_0
grpcio=1.46.3=py310h10ad472_0
gtk2=2.24.33=hbfd834a_1
gts=0.7.6=hde733a8_3
h5py=3.6.0=nompi_py310hb8bbf05_100
harfbuzz=2.8.1=hffeda63_0
hdf5=1.12.1=nompi_hf9525e8_104
icu=68.1=hc377ac9_0
idna=3.3=pyhd8ed1ab_0
joblib=1.1.0=pyhd3eb1b0_0
jpeg=9e=he4db4b2_2
keras=2.10.0=pypi_0
keras-preprocessing=1.1.2=pypi_0
kiwisolver=1.4.4=py310hd23d0e8_0
krb5=1.19.3=hf9b2bbe_0
lame=3.100=h1a8c8d9_1003
lcms2=2.12=had6a04f_0
lerc=4.0.0=h9a09cb3_0
libblas=3.9.0=16_osxarm64_openblas
libbrotlicommon=1.0.9=h1c322ee_7
libbrotlidec=1.0.9=h1c322ee_7
libbrotlienc=1.0.9=h1c322ee_7
libcblas=3.9.0=16_osxarm64_openblas
libclang=14.0.6=pypi_0
libcurl=7.85.0=hd538317_0
libcxx=14.0.6=h2692d47_0
libdeflate=1.14=h1a8c8d9_0
libedit=3.1.20191231=hc8eb9b7_2
libev=4.33=h642e427_1
libffi=3.4.2=h3422bc3_5
libgd=2.2.5=h6c9a55e_4
libgfortran=5.0.0=11_3_0_hd922786_25
libgfortran5=11.3.0=hdaf2cc0_25
libiconv=1.17=he4db4b2_0
libidn2=2.3.3=he4db4b2_0
liblapack=3.9.0=16_osxarm64_openblas
libnghttp2=1.47.0=h232270b_1
libopenblas=0.3.21=openmp_hc731615_3
libpng=1.6.38=h76d750c_0
libprotobuf=3.19.4=hccf11d3_0
librsvg=2.50.7=h90a7944_0
libsqlite=3.39.2=h2c9beb0_1
libssh2=1.10.0=hb80f160_3
libtasn1=4.19.0=h1a8c8d9_0
libtiff=4.4.0=hfa0b094_4
libtool=2.4.6=hc377ac9_1008
libunistring=0.9.10=h3422bc3_0
libvpx=1.11.0=hc470f4d_3
libwebp=1.2.4=h328b37c_0
libwebp-base=1.2.4=h57fd34a_0
libxcb=1.13=h9b22ae9_1004
libxml2=2.9.14=h8c5e841_0
libzlib=1.2.12=ha287fd2_2
llvm-openmp=14.0.4=hd125106_0
markdown=3.4.1=pypi_0
markupsafe=2.1.1=pypi_0
matplotlib=3.6.1=py310hb6292c7_0
matplotlib-base=3.6.1=py310h78c5c2f_0
munkres=1.1.4=pyh9f0ad1d_0
ncurses=6.3=h07bb92c_1
nettle=3.8.1=h63371fa_1
ninja=1.10.2=hca03da5_5
ninja-base=1.10.2=h525c30c_5
numpy=1.23.2=py310h127c7cf_0
oauthlib=3.2.2=pypi_0
openh264=2.3.1=hb7217d7_1
openjpeg=2.5.0=h5d4e404_1
openssl=1.1.1s=h03a7124_0
opt-einsum=3.3.0=pypi_0
p11-kit=0.24.1=h29577a5_0
packaging=21.3=pyhd8ed1ab_0
pango=1.48.7=h79d33a7_0
pcre=8.45=hc377ac9_0
pillow=9.2.0=py310h173adc3_2
pip=22.2.2=pyhd8ed1ab_0
pixman=0.40.0=h1a28f6b_0
protobuf=3.19.4=py310h1b49c16_0
pthread-stubs=0.4=h27ca646_1001
pyasn1=0.4.8=pypi_0
pyasn1-modules=0.2.8=pypi_0
pycosat=0.6.3=py310hf8d0d8f_1010
pycparser=2.21=pyhd8ed1ab_0
pydot=1.2.4=py_0
pyopenssl=22.0.0=pyhd8ed1ab_0
pyparsing=3.0.9=pyhd8ed1ab_0
pysocks=1.7.1=py310hbe9552e_5
python=3.10.6=hbce4517_0_cpython
python-dateutil=2.8.2=pyhd8ed1ab_0
python_abi=3.10=2_cp310
pytorch=1.12.1=py3.10_0
readline=8.1.2=h46ed386_0
requests=2.28.1=pyhd8ed1ab_0
requests-oauthlib=1.3.1=pypi_0
rsa=4.9=pypi_0
ruamel_yaml=0.15.80=py310h02f21da_1007
scikit-learn=1.0.2=py310h59830a0_1
scipy=1.9.1=py310ha0d8a01_0
setuptools=65.2.0=py310hbe9552e_0
six=1.16.0=pyh6c4a22f_0
svt-av1=1.2.1=he23bcac_0
tensorboard=2.10.1=pypi_0
tensorboard-data-server=0.6.1=pypi_0
tensorboard-plugin-wit=1.8.1=pypi_0
tensorflow-deps=2.10.0=0
tensorflow-estimator=2.10.0=pypi_0
tensorflow-macos=2.10.0=pypi_0
tensorflow-metal=0.6.0=pypi_0
termcolor=2.0.1=pypi_0
threadpoolctl=2.2.0=pyh0d69192_0
tk=8.6.12=he1e0b03_0
toolz=0.12.0=pyhd8ed1ab_0
torchvision=0.13.1=py310_cpu
tornado=6.2=py310h02f21da_0
tqdm=4.64.0=pyhd8ed1ab_0
typing_extensions=4.4.0=pyha770c72_0
tzdata=2022c=h191b570_0
unicodedata2=14.0.0=py310hf8d0d8f_1
urllib3=1.26.11=pyhd8ed1ab_0
werkzeug=2.2.2=pypi_0
wheel=0.37.1=pyhd8ed1ab_0
wrapt=1.14.1=pypi_0
x264=1!164.3095=h57fd34a_2
x265=3.5=hbc6ce65_3
xorg-libxau=1.0.9=h27ca646_0
xorg-libxdmcp=1.1.3=h27ca646_0
xz=5.2.6=h57fd34a_0
yaml=0.2.5=h3422bc3_2
zlib=1.2.12=ha287fd2_2
zstd=1.5.2=h8128057_4
